requirements.txt
flatten-dict
flywheel-sdk
-e https://github.com/flywheel-io/bids-client